Final Project/requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
cryptography==36.0.1
dateparser==1.1.0
decorator==5.1.0
defusedxml==0.7.1
Django==4.0
django-admin-interface==0.18.3
django-ckeditor==6.2.0
django-colorfield==0.5.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
greenlet==1.1.2
idna==3.3
ipython==7.30.1
jedi==0.18.1
jet-bridge-base==1.0.9
matplotlib-inline==0.1.3
oauthlib==3.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==2.0.9
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-core==4.1.0
SQLAlchemy==1.4.28
sqlparse==0.4.2
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wcwidth==0.2.5
